# ENH: optimisation of array_equal