--- a/numpy/core/numeric.py
+++ b/numpy/core/numeric.py
@@ -1,8 +1,10 @@
 def __getattr__(attr_name):
     from numpy._core import numeric
     from ._utils import _raise_warning
-    ret = getattr(numeric, attr_name, None)
-    if ret is None and attr_name != "newaxis":
+
+    sentinel = object()
+    ret = getattr(numeric, attr_name, sentinel)
+    if ret is sentinel:
         raise AttributeError(
             f"module 'numpy.core.numeric' has no attribute {attr_name}")
     _raise_warning(attr_name, "numeric")