For various reason I was looking at array_equal and I believe there are
a few case where we can get some speedup/less memory allocation.

  - if `equal_nan == True` then could we add a shortcut, we can sometime
    check that both array are the same and avoid the computation
  - if the dtypes cannot contain `nan`, then we do not need the expensive
    fancy indexing and removal of `nan`s

we can test will the following script

```python
import numpy as np

base = np.random.randint(1, 100, (10000, 10000), dtype='int64')
a = base.copy()
a1 = base.copy()

b = base.copy().astype('float')
b[b == 50] = np.nan

b1 = b.copy()

data = {'a':a, 'b':b, 'a1':a1, 'b1': b1}

for (x, y) in [('a','a'),('a','a1'),('b','b'),('b','b1')]:
    ax = data[x]
    bx = data[y]
    for en in [True, False]:
        res = np.array_equal(ax, bx, equal_nan=en)
        print(f"np.array_equal({x}, {y}, equal_nan={en}) == {res}")
        %timeit -r10 -o np.array_equal(ax, bx, equal_nan=en)

```

Which give us the following results.

```diff
--- before.txt  2023-09-07 15:19:01
+++ after.txt   2023-09-07 15:24:30
@@ -1,16 +1,16 @@
np.array_equal(a, a, equal_nan=True) == True
-419 ms ± 42.4 ms per loop (mean ± std. dev. of 10 runs, 1 loop each)
+1.41 µs ± 5.37 ns per loop (mean ± std. dev. of 10 runs, 1,000,000 loops each)

np.array_equal(a, a, equal_nan=False) == True
-31.1 ms ± 200 µs per loop (mean ± std. dev. of 10 runs, 10 loops each)
+31 ms ± 348 µs per loop (mean ± std. dev. of 10 runs, 10 loops each)

np.array_equal(a, a1, equal_nan=True) == True
-1.07 s ± 326 ms per loop (mean ± std. dev. of 10 runs, 1 loop each)
+40.3 ms ± 133 µs per loop (mean ± std. dev. of 10 runs, 10 loops each)

np.array_equal(a, a1, equal_nan=False) == True
-41.9 ms ± 130 µs per loop (mean ± std. dev. of 10 runs, 10 loops each)
+40.3 ms ± 210 µs per loop (mean ± std. dev. of 10 runs, 10 loops each)

p.array_equal(b, b, equal_nan=True) == True
-569 ms ± 61.7 ms per loop (mean ± std. dev. of 10 runs, 1 loop each)
+909 ns ± 1.61 ns per loop (mean ± std. dev. of 10 runs, 1,000,000 loops each)

np.array_equal(b, b, equal_nan=False) == False
-31.2 ms ± 208 µs per loop (mean ± std. dev. of 10 runs, 10 loops each)
+30.2 ms ± 228 µs per loop (mean ± std. dev. of 10 runs, 10 loops each)

np.array_equal(b, b1, equal_nan=True) == True
-842 ms ± 171 ms per loop (mean ± std. dev. of 10 runs, 1 loop each)
+461 ms ± 3.07 ms per loop (mean ± std. dev. of 10 runs, 1 loop each)

np.array_equal(b, b1, equal_nan=False) == False
-40.4 ms ± 383 µs per loop (mean ± std. dev. of 10 runs, 10 loops each)
+39.2 ms ± 234 µs per loop (mean ± std. dev. of 10 runs, 10 loops each)
```