@seberg 2023-09-07
Just my first thoughts:  I am not sure I think it needs to be our job to check for identity, even when we know its correct.  I am happy to skip NaN checks when we know for sure that there are no NaNs.
One thing to beware of: timedeltas have NaNs, and I am not quite sure you are catching them (which I agree is wrong by NumPy if the case, but we need to take care here either way).