@Carreau 2023-09-07
> I am not sure I think it needs to be our job to check for identity, even when we know its correct

Yeah, I'm unsure, I think it's pretty rare, but at the same time, it can be quite a bit more performant.


> One thing to beware of: timedeltas have NaNs, and I am not quite sure you are catching them (which I agree is wrong by NumPy if the case, but we need to take care here either way).

Do you mean we should have an explicit list of non-nanable types instead of issubtype ? 

Regardless both are not a hill I'm going to die on, I'm happy to split the logic in two PRs if you wish.