@seberg 2023-09-07
I am just not sure that your current `issubdtype` (which I don't like, but there isn't a much cleaner alternative probably) doesn't also match for timedeltas: this needs a test.