@seberg 2023-09-07
Actually an alternative that we may or may not like:  I always think that it might be useful to add a ufunc for [`islessgreater`](https://en.cppreference.com/w/c/numeric/math/islessgreater) or maybe an inverted version.  Maybe with a better name `not_different`?!

I.e. so that we don't have to do the NaN dance at all, but have a ufunc that matches NaNs (I think this would also help fix a bunch of `np.unique` annoyances/bugs).