@Carreau 2023-09-08
The approach `issubdtype` does indeed fail for timestamp64, I decided to fall back to a `dtype in {...}` with an explicit list of dtype that cannot have NaNs, in a utility function that would make it easy to update this list. 

I've refactored the test to use pytest parametrize, and added a number of case of interest, and rebased to put the tests commit first (modified test is passing before the PR).