@Carreau 2023-09-10
Added a test for 0-d arrays in the first commit. 

I can remove the `asarray(? == ?)` in the function but I'm not entirely confident about it.