@Carreau 2023-09-11
> .e. so that we don't have to do the NaN dance at all, but have a ufunc that matches NaNs (I think this would also help fix a bunch of `np.unique` annoyances/bugs

Maybe stupid question, I'ts not my area of expertice but may assuming array have same dtypes, woudln't `a.tobytes() == b.tobytes()` work for this ? I guess `tobytes()` make copies, but maybe there is a way to do it w/o copies ?