@Carreau 2023-09-11
> Maybe stupid question, I'ts not my area of expertice but may assuming array have same dtypes, woudln't `a.tobytes() == b.tobytes()` work for this ? I guess `tobytes()` make copies, but maybe there is a way to do it w/o copies ?

Just tests, that does not work for `1+nan.j == nan + 1.j`, but I guess we have at least `a.tobytes() == b.tobytes()` =>  `a == b`. Maybe that's an extra optimisation.