@ngoldbaum 2023-09-11
> woudln't a.tobytes() == b.tobytes() work for this 

I think it might work for all but object arrays since you can't know the truthiness of a python value without testing it.