@Carreau 2023-09-21
For both the `is` question and the `dtype` micro-optimisation I think depends on wether you want to optimize for large or small arrays.

I'm tempted to want to optimize for large array (even if my use case is small-ish). This is were we'll see the most gain. Timing wise as we to at least µs, a few ns for dtype checking won't make much the differences. For large array checking, the `is` can have also a large impact on memory allocation.