@Carreau 2023-10-02
> There is still the question about the complex test that @rossbar wrote where I'm not sure what was written was what was intended.

Updated comments following discussion.