@Carreau 2023-10-26
Gentle nudge again, as this was relatively painful to rebase on master after the recent changes. 

I can again pull the test refactor into its own PR if that makes it easier to get accepted.