@Carreau 2023-10-26
> It looks like some of the new tests are failing.

Yeah, I've botched the merge/rebase. Redoing now.

> If you'd like, stop by the next NumPy triage zoom meeting on Wednesday at 1:00 PM US Eastern time and you'll have more people's attention.

I'll try, but Those times are usually touch for me here, and today I'll can't for sure.

I've pulled out the test refactor into its own PR #25010