@Carreau 2023-10-30
Here are some redone bench test with 1000x1000 random array: 
![BENCH](https://github.com/numpy/numpy/assets/335567/aef4e01d-3154-466f-a33d-c13cfbc0d8c7)

It's only a guideline as the larger the arrays are the more the effect will be felt.

(Source to replicate : https://gist.github.com/Carreau/baecfc0b02773bea6a744935e0dc9d1b)

CI seem to have a problem with np.dtypes, but it works locally, which is weird.