@seberg 2024-08-23
Just to note in case anyone has time.  We need to tweak this a bit and add a regression test.  This unfortunately broke some comparisons of scalar arrays with `array_equal`.

(I think the reason is only historic special paths that lead to real booleans rather than NumPy ones... but what can you do...)

See gh-27271