@eendebakpt 2024-08-23
@seberg Adding back the `asarray` would probably fix this. I assume you do not want to attempt the other option: update the code so that `np.array("foo") == np.array(1)` returns a numpy bool?