@eendebakpt 2024-08-23
I created https://github.com/numpy/numpy/pull/27275